avoid accidental 32-bit truncation of positions in paired end MEM chainer

+++ src/cluster.cpp
@@ -15,7 +15,7 @@ using namespace std;
 MEMChainModel::MEMChainModel(
     const vector<size_t>& aln_lengths,
     const vector<vector<MaximalExactMatch> >& matches,
-    const function<int(pos_t)>& approx_position,
+    const function<int64_t(pos_t)>& approx_position,
     const function<double(const MaximalExactMatch&, const MaximalExactMatch&)>& transition_weight,
     int band_width,
     int position_depth,
