+++ src/cluster.hpp
@@ -133,7 +133,7 @@ public:
     MEMChainModel(
         const vector<size_t>& aln_lengths,
         const vector<vector<MaximalExactMatch> >& matches,
-        const function<int(pos_t)>& approx_position,
+        const function<int64_t(pos_t)>& approx_position,
         const function<double(const MaximalExactMatch&, const MaximalExactMatch&)>& transition_weight,
         int band_width = 10,
         int position_depth = 1,
